move arithmetic rewriting tests
